requirements.txt
langchain
langchain_groq
langchain-community
langchain-huggingface
chromadb
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
python-dotenv
sentence-transformers
# starlette -> REMOVE this line (FastAPI will install a compatible one)
pandas>=2.0
